requirements.txt
pdf2image
python-dotenv
google-cloud-aiplatform
google-cloud-documentai
langchain
langchain-community
langchain-core
langchain_google_genai
sentence-transformers
torch
langchain_chroma
langchain_google_vertexai
google_crc32c